+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Make account lookup by username and email address case insensitive. See [`UPGRADING.md`](./UPGRADING.md) for details. (COMPASS-333)
+
 ## [4.2.2] - 2026-08-20
 
 ### Fixed